pyproject.toml
[tool.poetry]
name = "ecommerce-rest"
version = "0.1.0"
description = ""
authors = ["Gerson Vizquel <aituring.app@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django = "^4.2.7"
django-simple-history = "^3.4.0"
pillow = "^10.1.0"
numpy = "^1.26.2"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.1.0"
django-extensions = "^3.2.3"
deptry = "^0.12.0"
djlint = "^1.31.1"
pre-commit = "^3.3.3"
pygments = "^2.16.1"
ruff = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 90
ignore=["E501"]
target-version = "py310"
indent-width = 4

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
